[FEATURE] add program content exercise block
